+++ engine/src/main/scala/cromwell/engine/workflow/lifecycle/execution/callcaching/CallCacheReadActor.scala
@@ -1,10 +1,12 @@
 package cromwell.engine.workflow.lifecycle.execution.callcaching
 
-import akka.actor.{Actor, ActorLogging, ActorRef, Props}
+import akka.actor.{Actor, ActorLogging, ActorRef, Props, Status}
 import akka.pattern.pipe
 import cats.data.NonEmptyList
 import cromwell.core.Dispatcher.EngineDispatcher
 import cromwell.core.callcaching.HashResult
+import cromwell.database.sql.joins.CallCachingDiffJoin
+import cromwell.database.sql.tables.CallCachingEntry
 import cromwell.engine.workflow.lifecycle.execution.callcaching.CallCacheReadActor._
 
 import scala.concurrent.ExecutionContext
@@ -25,6 +27,9 @@ class CallCacheReadActor(cache: CallCache) extends Actor with ActorLogging {
     case request: CallCacheReadActorRequest => receiveNewRequest(request)
     case response: CallCacheReadActorResponse =>
       currentRequester foreach { _ ! response }
+      cycleRequestQueue()
+    case Status.Failure(f) =>
+      currentRequester foreach { _ ! CacheResultLookupFailure(new Exception(s"Call Cache query failure: ${f.getMessage}")) }
       cycleRequestQueue()
     case other =>
       log.error("Unexpected message type to CallCacheReadActor: " + other.getClass.getSimpleName)
@@ -47,8 +52,22 @@ class CallCacheReadActor(cache: CallCache) extends Actor with ActorLogging {
           case Some(nextHit) => CacheLookupNextHit(nextHit)
           case None => CacheLookupNoHit
         }
+      case CallCacheDiffRequest(parameters) => cache.callCacheDiff(parameters.callA, parameters.callB) map {
+        case CallCachingDiffJoin(cacheEntryA, cacheEntryB, diff) =>
+          val formatted = diff flatMap {
+            // If both are Some we assume the keys are the same, as they should be
+            case (maybeKeyValueA, maybeKeyValueB) =>
+              val maybeKey = maybeKeyValueA.map(_._1).orElse(maybeKeyValueB.map(_._1))
+              val maybeValueA = maybeKeyValueA.map(_._2)
+              val maybeValueB = maybeKeyValueB.map(_._2)
+
+              maybeKey map { CallCachingDiffElement(_, maybeValueA, maybeValueB) }
+          }
+
+          CallCachingDiff(cacheEntryA, cacheEntryB, formatted)
+      }
     }
-    
+
     val recovered = response recover {
       case t => CacheResultLookupFailure(t)
     }
@@ -67,7 +86,7 @@ class CallCacheReadActor(cache: CallCache) extends Actor with ActorLogging {
   }
 
   private def receiveNewRequest(request: CallCacheReadActorRequest): Unit = currentRequester match {
-    case Some(x) => requestQueue :+= RequestTuple(sender, request)
+    case Some(_) => requestQueue :+= RequestTuple(sender, request)
     case None =>
       currentRequester = Option(sender)
       runRequest(request)
@@ -90,7 +109,8 @@ object CallCacheReadActor {
   case class CacheLookupRequest(aggregatedCallHashes: AggregatedCallHashes, cacheHitNumber: Int) extends CallCacheReadActorRequest
   case class HasMatchingInitialHashLookup(aggregatedTaskHash: String) extends CallCacheReadActorRequest
   case class HasMatchingInputFilesHashLookup(fileHashes: NonEmptyList[HashResult]) extends CallCacheReadActorRequest
-  
+  case class CallCacheDiffRequest(queryParameter: CallCacheDiffQueryParameter) extends CallCacheReadActorRequest
+
   sealed trait CallCacheReadActorResponse
   // Responses on whether or not there is at least one matching entry (can for initial matches of file matches)
   case object HasMatchingEntries extends CallCacheReadActorResponse
@@ -99,7 +119,14 @@ object CallCacheReadActor {
   // Responses when asking for the next cache hit
   case class CacheLookupNextHit(hit: CallCachingEntryId) extends CallCacheReadActorResponse
   case object CacheLookupNoHit extends CallCacheReadActorResponse
-  
+
+  // Response for call cache diff
+  case class CallCachingDiffElement(hashKey: String, hashValueA: Option[String], hashValueB: Option[String])
+  case class CallCachingDiff(cacheEntryA: CallCachingEntry,
+                             cacheEntryB: CallCachingEntry,
+                             hashDifferential: Seq[CallCachingDiffElement]
+                            ) extends CallCacheReadActorResponse
+
   // Failure Response
   case class CacheResultLookupFailure(reason: Throwable) extends CallCacheReadActorResponse
 }