+++ engine/src/main/scala/cromwell/webservice/WorkflowJsonSupport.scala
@@ -10,7 +10,9 @@ import MetadataService.{WorkflowQueryResponse, WorkflowQueryResult}
 import cromwell.util.JsonFormatting.WdlValueJsonFormatter
 import WdlValueJsonFormatter._
 import better.files.File
-import spray.json.{DefaultJsonProtocol, JsString, JsValue, RootJsonFormat}
+import cromwell.database.sql.tables.CallCachingEntry
+import cromwell.engine.workflow.lifecycle.execution.callcaching.CallCacheReadActor.{CallCachingDiff, CallCachingDiffElement}
+import spray.json._
 
 object WorkflowJsonSupport extends DefaultJsonProtocol {
   implicit val workflowStatusResponseProtocol = jsonFormat2(WorkflowStatusResponse)
@@ -28,6 +30,37 @@ object WorkflowJsonSupport extends DefaultJsonProtocol {
       case JsString(str) => Paths.get(str)
       case unknown => throw new NotImplementedError(s"Cannot parse $unknown to a File")
     }
+  }
+
+
+  implicit object callCachingDiffElementJsonFormat extends RootJsonFormat[CallCachingDiffElement] {
+    override def write(obj: CallCachingDiffElement) =
+      JsObject(Map(
+        obj.hashKey -> JsObject(Map(
+          "callA" -> obj.hashValueA.toJson,
+          "callB" -> obj.hashValueB.toJson
+        ))
+      ))
+    override def read(json: JsValue): CallCachingDiffElement =
+      throw new NotImplementedError(s"Cannot parse json to CallCachingDiffElement")
+  }
+  
+  implicit object callCachingDiffJsonFormat extends RootJsonFormat[CallCachingDiff] {
+    def makeCallObject(cacheEntry: CallCachingEntry) =  JsObject(Map(
+      "workflowId" -> JsString(cacheEntry.workflowExecutionUuid),
+      "callFqn" -> JsString(cacheEntry.callFullyQualifiedName),
+      "jobIndex" -> JsNumber(cacheEntry.jobIndex),
+      "allowResultReuse" -> JsBoolean(cacheEntry.allowResultReuse)
+    ))
+    
+    override def write(obj: CallCachingDiff) = 
+      JsObject(Map(
+        "callA" -> makeCallObject(obj.cacheEntryA),
+        "callB" -> makeCallObject(obj.cacheEntryB),
+        "hashDifferential" -> obj.hashDifferential.toList.toJson
+      ))
+    override def read(json: JsValue): CallCachingDiff = 
+      throw new NotImplementedError(s"Cannot parse json to CallCachingDiff")
   }
 
   implicit val workflowSourceDataWithImports = jsonFormat5(WorkflowSourceFilesWithDependenciesZip)